Split lexer tests
